Re-architecture Phase 2: promote EntitySourceDirectory to top-level class
Extraction-doc step 7: EntitySourceDirectory leaves ProjectStructure as a
top-level class in the same package, taking its two private path-append
helpers with it. Refactoring only; no behavior change.
