+++ legend-sdlc-server/src/test/java/org/finos/legend/sdlc/server/project/TestEntitySourceDirectory.java
@@ -19,7 +19,6 @@ import org.eclipse.collections.api.list.ImmutableList;
 import org.finos.legend.sdlc.domain.model.entity.Entity;
 import org.finos.legend.sdlc.serialization.EntitySerializers;
 import org.finos.legend.sdlc.serialization.EntityTextSerializer;
-import org.finos.legend.sdlc.server.project.ProjectStructure.EntitySourceDirectory;
 import org.junit.Assert;
 import org.junit.Test;
